Report errors caught while trying to get enketo edit url
These are mostly 400 level errors.  We ask the user to contact support
and so keeping a record of the errors encountered is a good idea, for
referencing.

In the spirit of DRY, we merged the duplicated report_exception functions
in logger_tools and viewer_tools and moved it to common_tools.
