+++ onadata/libs/utils/common_tools.py
@@ -1,5 +1,13 @@
 import six
 import uuid
+import sys
+import traceback
+
+from django.core.mail import mail_admins
+from django.utils.translation import ugettext as _
+from django.conf import settings
+
+from raven.contrib.django.raven_compat.models import client
 
 
 def str_to_bool(s):
@@ -27,3 +35,30 @@ def getUUID():
     Return UUID
     '''
     return uuid.uuid4().hex
+
+
+def report_exception(subject, info, exc_info=None):
+    # Add hostname to subject mail
+
+    subject = "{0} - {1}".format(subject, settings.HOSTNAME)
+    if exc_info:
+        cls, err = exc_info[:2]
+        message = _(u"Exception in request:"
+                    u" %(class)s: %(error)s")\
+            % {'class': cls.__name__, 'error': err}
+        message += u"".join(traceback.format_exception(*exc_info))
+
+        # send to sentry
+        try:
+            client.captureException(exc_info)
+        except Exception:
+            # fail silently
+            pass
+    else:
+        message = u"%s" % info
+
+    if settings.DEBUG or settings.TESTING_MODE:
+        sys.stdout.write("Subject: %s\n" % subject)
+        sys.stdout.write("Message: %s\n" % message)
+    else:
+        mail_admins(subject=subject, message=message)
